integrator tests behave like a proper mixin now
